requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.8 # Apache-2.0

click
PyYAML>=3.10.0
requests>=2.10.0,!=2.12.2 
